move FXModel to imx.py
